Prep for Json reading/Complex rendering
Remove MaterialPrefab, won't be a concept we explore (I provide materials
that can be used for comparisons to check if a block is a custom content).

Add constructor to block that takes a gson object (Google Json, Minecraft
uses it) which has the block characteristics.

TODO: We can now let an addon control the block's textures, need to do so

Signed-off-by: Chris Sanders <zidane@outlook.com>